--code: Sort CapDL objects secondarily on their name.
This makes it possible, given suitable alignment of the stars, for someone
supplying a CapDL specification to the translator to end up with physically
contiguous frames. Specifically, this allows CAmkES to usually arrive at a
physically contiguous DMA frame pool. This is not especially principled, and
users shouldn't rely on it being bulletproof, but it solves an immediate
practical problem for CAmkES users.

Closes JIRA CAMKES-193