Fix: Do not assume device frames are 4K.
This commit adds an extra parameter to `find_device_frame` to specify the size
of the frame. This has no effect on the 'master' kernel, other than permitting
some extra sanity checks. On the 'experimental' kernel, it removes an
assumption that all requested device frames are 4K. This assumption is not true
for all devices and has caused problems for the Kitty v4 application.

Closes JIRA CAMKES-352