Sort a spec on printing.
This resolves an issue with an external caching tool. The tool operates by
reading the input (in this case a textual CapDL spec) to a compiler or similar,
taking its hash and comparing against the hashes of previous inputs it has
seen. If it finds it has seen this input before, it can retrieve a previously
cached output and save running the compiler. This works rather like Ccache.

Unfortunately this caching becomes more-or-less useless when semantically
meaningless permutations happen in the input. In this case, the specification
coming out in a slightly different order (due to Python's string hashes not
being constant across executions) was causing unnecessary cache misses.

Fixes GINCA #9