Expose a function for calculating CNode sizes.
This function is primarily useful in the CAmkES template context, where it can
used to inspect or predict the eventual size of a CNode. Note that using this
function on generated CNodes from within a CAmkES template will only work if
there is no more cap allocation to be done. That is, if you call this function
from a connection template, some further capabilities may get allocated *after*
you execute, leading the final CNode size to be greater than the value you have
obtained.