Set fault endpoints correctly.
Prior to this commit, the CapDL initialiser would expect TCBs' fault endpoints
to be a capability, addressed in the same way as a CNode or a PD when
configuring a TCB. This is not how the seL4 API functions, and in fact an index
into the *target TCB's* CNode is expected here. The result of this incongruity
was that the initialiser used an index into *its own* CNode when configuring
TCBs. The fault endpoint slot would be invalid at runtime when the thread
faulted and a cap fault would be triggered, as if the thread had no configured
fault endpoint.

In unfortunate cases, the CSlot index of the capability in the initialiser's
CSpace would collide with the CSlot index of another endpoint in the target
thread's CNode. The result of this would be a cap fault in the target thread
causing an unexpected IPC. In a CAmkES system, this would be detected as a
corrupted or malicious RPC message that would trigger error handling code that
was not anticipating coping with this particular scenario and would
unnecessarily halt the entire system. Not pleasant.

Closes JIRA VER-348