Remove weakref caching in page collections.
This caching was intended to reduce memory usage in high loads by dropping the
cached CapDL specification under pressure. In practice, this is never a concern
and introduces unnecessarily complicated code in this functionality. This
commit leaves the caching in place, but removes the use of weak references.
Users are unlikely to perceive any difference in either memory usage or
runtime.