Allow frames to be allocated with an indicated size.
Frames are generally allocated as either `seL4_FrameObject` or one of the
architecture-specific 4KB or large frame types. However, in template code it is
often much more convenient and portable to just allocate all of these as
generic frames with a given size. Where previously you would need to write
something like the following in the CAmkES template context:

    /*- if arch == 'arm' and size == 64 * 1024-*/
      /*- set f = alloc('foo', seL4_ARM_LargePageObject) -*/
    /*- elif arch == 'ia32' and size == 4 * 1024 * 1024 -*/
      /*- set f = alloc('foo', seL4_IA32_4M) -*/
    ...

you can now write:

    /*- set f = alloc('foo', seL4_FrameObject, size=size) -*/

Note that it was possible previously to do this on an individual object basis,
but this exposes such flexibility in the object allocator as well.