optimisation: Remove indentation in C code output.
Due to the way in which the C code output is constructed, the indentation
operation is quite expensive. On a sample 41MB CapDL textual spec when running:

    parse-capDL --code /dev/null sample.cdl

this reduces processing time by ~12% and memory usage marginally. Though these
gains are modest, this is most likely just the tip of low hanging optimisation
fruit in this backend.