C backend: Remove `driverinfo` struct member.
This commit removes explicit assignment to the `driverinfo` member of the TCB
struct when emitting the CapDL spec. This currently has no effect on the
generated struct because it was always being set to 0 and this field defaults
to zero anyway. However, this change makes it possible to remove the
`driverinfo` from the struct definition (in the CapDL initialiser) if/when we
decide it is no longer worth maintaining this bitrotted data.