Allow a cap in a cnode to specify mapping information
This addresses the case where a thread needs a frame cap with an
associated mapping. An example of such a case is when a thread needs to
flush a frame's contents from the cpu cache. To do this, the cap that
was used to establish the relevant mapping must be invoked (invoking a
copy will not work in this case). To support this behaviour, capdl needs
a way of specifying the association of a cap in a cnode to a "cap" in a
slot in a page table/directory specifying a mapping.