python-capdl: Change some `filter` calls to list comprehensions.
This commit converts two calls to `filter` into the corresponding list
comprehension. The motivation for this is movement towards Python 3
compatibility. In Python 3, `filter` returns a generator, not a list and so a
result that is expected to be iterated over multiple times does not function as
expected. There are many more changes needed for full Python 3 compatibility.
Python 2 functionality should be unaffected.