capDL-tool: Make Isabelle output loadable from any directory.
Prior to this commit, the Isabelle back end for CapDL specs assumes that the
output theory, CapDLSpec.thy, will be hosted in l4v/spec/capDL, the same
directory as Types_D.thy. This commit makes the import path to Types_D relative
to the Isabelle source directory (for which `~~` is a shortcut) itself. This
makes it usable from anywhere in your file system when using Isabelle from an
l4.verified checkout.