python-capdl: Remove now unused import.
This fixes a circular dependency within this package. This change is motivated
by Python 3 compatibility. Everything worked fine previously under Python 2, but
Python 3 requires these imports to be relative and then fails on this cycle in
the dependencies here. The issue is that a dependency like this should never be
allowed, but Python 2 accidentally permits this. Following this change, we are
still not completely compatible with Python 3, but getting there bit-by-bit.