python-capdl: Omit TCBs' fault_ep_slot when not present.
This has no effect on systems booted by the CapDL initialiser, as 0 is the
default fault_ep_slot. The purpose of this change is that, when set, the
fault_ep_slot is translated into the Isabelle representation of CapDL as
setting the cdl_tcb_has_fault member of the TCB's record. This causes flow on
problems when trying to prove correspondence to the output of the CAmkES
generator function.

This cherry picks 8b2fcbe68bca83839b04442fbe29543ffa3fe5c6 from
GINCA:mfernandez/project.