python-capdl: Use normalise_architecture for increased ARM Hyp mode compat.
This commit allows the util function to recognise the architecture string
"arm_hyp" as a way of indicating ARM's Hyp mode. That is, you can now specify
Hyp mode in two ways:

  1. architecture "arm" and hyp=True (legacy); or
  2. architecture "arm_hyp".

The purpose of this commit is movement towards removing the `hyp` argument
entirely and having a single, consolidated way of indicating the target
architecture. Note that none of this new functionality is currently used yet.