CapDL init: x86: Change calling convention when passing arguments in registers.
** BREAKING CHANGE **

The CapDL initialiser has two options for passing arguments to threads it is
starting: the standard calling convention (cdecl) or passing arguments entirely
in registers. The default is the former. This commit modifies the latter
convention on x86 to pass arguments from left to right in the registers ECX,
EDX, EAX, EBX. The previous convention was EAX, EBX, ECX, EDX.

This change should be irrelevant for most users, but it has an important
benefit in that for cases where two or fewer arguments are passed, the
convention now aligns with the fastcall convention. This means, for a limited
use case, it is now possible to write your thread's entry point in C:

    void _start(unsigned long arg0, unsigned long arg1)
      __attribute__((fastcall, noreturn));

This is marked as a breaking change because threads that are started by the
CapDL initialiser using this calling convention on x86 will need to be updated
to expect their arguments in different registers. This change has no effect for
ARM applications or threads that expect their arguments to be provided via the
standard calling convention.