python-capdl-tool: Define a real architecture definition
This replaces and expands upon the architecture abstractions in util.py.
Previously the architecture just abstracted some sizes of pages and
layouts of a two level paging hierarchy. Now this is generalized both
for arbitrary vspace hierarchies, as well as other architecture
information.

The original interface is left and emulated.