Find copy addr frame cap using lowest mapped vaddr
The capdl loader copies data from elf images into memory by mapping
frames to back a global array. Before this can happen, the frame
originally backing this array must be unmapped, and thus its cap must be
identified in the list of user image frames in bootinfo. This was
previously done under the assumption that the first frame in the user
image corresponded to the __executable_start symbol, but this is not
necessarily the case. This change uses the number of user image frames
and the difference between the __executable_start and _end symbols to
determine the vaddr corresponding to the first frame of the user image,
which can be used to identify the desired frame cap.