Remove driverinfo support.
This data was previously used by a branch of the CapDL tools that has long been
deprecated. I think it's safe to say at this point that it won't be relevant
any time soon. The motivation for this commit is reducing in-memory (and
in-image for load) CapDL spec sizes. This commit reduces the ARM VM image by
64KB.