Reduce size_bits and pack the CDL_Object struct.
The `size_bits` of an object can, at most, be 32. In a practical system it can
be at most 27. By reducing the size of this member, moving it to the end of the
struct and packing the struct we reduce the overall size/alignment of the
struct. This, again, reduces the in-memory size of CapDL specs. This reduces
the size of the ARM VM by 55KB.