Make domain a uint8_t.
Unlike the previous commits, this has no effect on the size of a CapDL spec
in-memory. The motivation for this commit is consistency. The API to the kernel
only gives you an 8-bit wide value for specifying the domain of a thread, so
representing a domain as a 32-bit value in CapDL falsely gives users the
impression that they can setup more than 256 domains.