init: Fix: Consider reply caps as EP caps.
This commit introduces a function `ep_related_cap` to match the existing
definition in l4.verified [0]. In doing so, it fixes a problem whereby reply
caps were not considered to be endpoint caps. This is unlikely to have affected
any users as there's no rational use case for initialising a system with reply
caps.

  [0]: https://github.com/seL4/l4v/blob/7e9b8224ee95874c517cbd5c9f28209d01c00360/spec/capDL/Syscall_D.thy#L86-L91