init: Remove verification configuration flag.
The initialiser is never regularly tested with this flag and code has crept in
that is not in our supported C subset (e.g. C bitfields). Furthermore, because
this configuration combination is never tested, it also contains inconsistent
paths (e.g. a partial understanding of MSI etc interrupts). This commit removes
the flag entirely, reducing the number of configuration permutations that need
to be mentally managed when developing on this code base. When (if) code-level
verification is re-introduced in future, a pass through the l4.verified model to
re-align this code can easily pull it back into shape.