Limited support for frames with overlapping addresses
After commit c4b9c519a403324 it was no longer possible for a capDL
spec to request two device frames with the same physical address. This
limitation was due to how the device untypeds were processed and turned
into frames.

This commit partially fixes this by looking at the previous allocation to
see if it perfectly overlaps with the new object, and if so duplicates
the capability.