capDL: Support 'fill' parameter on frame objects
Adds a 'fill' paramter on frame objects that is an arbitrary description on the initial
state of contents of this frame that an initializer should create. Arbitrary means that
the description is designed to be extensible and largely pass information from one
layer to another. The two defined pieces of information are a 'type' (which is a string)
and an integer offset into the frame. Any other words of information get turned into a
string without looking at them for another layer (such as the capDL loader) to parse.