Capdl loader only prints using seL4_DebugPutChar
Previously, if kernel debugging was disabled (preventing the use of
seL4_DebugPutChar), the capdl loader would attempt to initialize a
serial driver to use for printing. This involved retyping memory using
allocman, the effects of which went untracked by the capdl loader.

This change prevents the loader from initializing a serial driver,
preventing printing when kernel debugging is disabled.