Re-add missing '-- Dump ' to the start of test
This causes the capdl translator to treat the file as a "dump" rather
than a regular capdl spec.

Note the trailing whitespace in '-- Dump '.

This cdl syntax looks like a haskell comment, but it's not a cdl comment
(those are in /* ... */), so it was removed as part of a recent
refactor.