capdl-loader-app: cmake build support
Provides cmake build support for the capdl-loader. As the capdl-loader is not a standalone
application, but rather requires configuration from the surrounding system, it is implemented
partially through a helper function through which the user/system provides the capdl
specification.

The choice of having the add_executable performed in the helper, with configuration for the
static files coming through target properties, versus providing the capdl spec location
through target properties is largely arbitrary although there are a couple of small
advantages to the current setup
 * Files provided through target properties need to be given by absolute path as the
   current source/build directory will be different between the app CMakeLists and
   the helper function and it is easier to do absolute paths for the known static C
   files than compute arbitrary paths for user provided locations that could be relative
   paths to nonexistent files and we have to guess where they are
 * Provides the add_executable only if and where the system wants it, which makes a bit more
   sense conceptually