Update IOPort model
Changes the model of the IOPorts in the capdl specification and the surrounding tools
to reflect a change in the kernel model. There is now a single IO port control capability
that can produce non overlapping IO port caps. Due to the non overlapping requirement the
capdl model treats the 'object' as an IO port cap issued from the control capability and
this is then duplicated at the cap level to produce all cap references.