capDL-tool: sort objects in Isabelle output
This refactors the sorting code so that PrintC and PrintIsabelle use the
same object ordering. The camkes-tool's capDL spec and generated proofs
will also use the same ordering, and hopefully this consistency will
lead to fewer unpleasant surprises in the future.

Adapted from a similar refactoring in Matthew Fernandez's own repo.