python-capdl: Add AddressSpaceAllocator class
Similar to ObjectAllocator and CSpaceAllocator this class is for
tracking allocations as a CapDL spec is constructed. This class is for
tracking the allocation of special backing frames for a single address
space. Symbols that have a greater meaning than anonymous memory such as
IPC buffers, shared memory, guarded symbols, unmapped symbols, DMA pools,
Physical memory mappings, etc should be added to this with structure
with Caps referring to frames that have already been allocated with the
desired attributes. When the final address space is created from the ELF
file, this structure will be passed to the PageCollection object which
will try and use the Caps that have already been created instead of
creating new Frames. This should help applications like Camkes to more
easily manage special frames as they can be entirely handled in the same
location that the symbols are defined with no additional book keeping.