capdl: allow IRQControl to be in loaded cnode
This commit allows the IRQControl capability to be loaded by the capdl
loader into a single cspace (you cannot create copies of this cap, so
trying to put it into two cspaces will fail).

There are no changes to the Haskell, which already supports irq_control.

Changes to python:

- change ObjectType.seL4_IRQControl to ObjectType.IRQHandler. This looks
like an error in the existing code, as the constant was always
interpreted as an IRQHandler not the IRQControl capability.
- add a new ObjectType.seL4_IRQControl
- add a new Object.IRQControl class which resolves to "irq_control".
This matches how other control capabilities are treated in the python,
although they do not actually resolve to Objects (just caps).

Changes to C:

- add a new capdl_to_irq_ctrl data structure which mirrors the
sched_ctrl data structure
- when irq_control is specified in a cnode, move it to that slot