python-capdl-tool: add object sizes
This change allows a dictionary of object sizes to be registered with
the python capdl tool (via Object.register_object_sizes) which can then
be queried to find the size of non-fixed sized object types.

The dictionary provided should define sizes for all fixed-sized objects
in the Object.ObjectType enum that are used by the spec being processed
with the library.