python-capdl-tool: add untyped allocator
This change adds an offline allocator for calculating untyped
derivations, which will be used in future to specify exact derivations
from boot info untyped objects. This allows the allocation logic to be
removed from the loader, this simplifying the loader model for
verificaion while allowing for policy freedom in terms of allocation.

The initial allocator implementation is a best-fit allocator.