capdl-loader-app: Change type from byte to word
In this struct, there were no other byte sizes values to pack
in along with the type field, so the struct size is not altered by using
a uint8 vs a word. This change makes the code clearer and faster, as the
compiler no longer has to pack and unpack the byte value.