python-capdl-tool: Add AllocatorState from camkes
Contains allocation data of objects in a CapDL build. There is currently
a pattern of serialising the different capdl allocators
(ObjectAllocator, CSpaceAllocator, AddressSpaceAllocator) between
allocating shared caps and objects which occurs before component
binaries are built, and creating the remaining objects inferred from the
resulting ELF files, which occurs after building. We want to use
AllocatorState from Camkes as a common type to serialise that the
capdl-ld script will use.