capdl_linker: Change file data types
The capdl_linker tool has two modes: build_cnode and gen_capdl. Both
were used by the seL4 tutorials and so the input configuration file was
simply a python serialised blob that was valid for both modes. We now
split this into a different file format for each of the two stages. This
allows us to use a more human readable YAML format for build_cnode, and
for gen_capdl are format that can be used by both the tutorials and
camkes.