Add keys argument and use as label
This is to avoid inferring the key for objects in AllocatorState based
on the elf file name. Labels are important because they are used for
integrity refinement proofs. The keys passed in must match with whatever
labels were used to allocate any existing objects in the AllocatorState
as any new objects created from ELF files will use the keys as the label
name.