python-capdl: Allow page_sizes to take Arch type
If page_sizes is called with an actual Arch object type (or an object
that implements the get_pages method, then don't lookup the architecture
again. This is to support cases where an Arch object instance is
available, but the string that was used to obtain it is not in scope.