capdl-loader-app: Remove 'simple' from init_frame
This commit refactors the copying of the extended bootinfo information
from init_frame. We add a function which looks up the extended bootinfo
header table and copies chunks of the extended bootinfo information
across to another buffer. This is a step towards exposing the DTB to
CAmkES userland and allowing the fill frames mechanism to interact with
non-4K page-sized extended bootinfo blocks.