capdl-loader-app: remove superfluous page flush
The loader sets aside a 4K page for init_frame. init_copy_frame unmaps
this page, so that init_frame can map user frames into it later.
Prior to unmapping this page, the loader would flush it; this flush
invocation is being removed.

There are both endogeneous and exogeneous reasons for this removal:
 - The loader's page is not being used for anything, and furthermore,
   it is a data page, not code, so it makes no sense to flush this page
   *from the I-cache* before unmapping it.

 - Trying to flush this page on some AARCH32 hypervisor platforms would
   in fact crash the kernel. This can happen when the loader is
   ELF-loaded to physical memory that is outside the kernel virtual
   window. (This is now bug SELFOUR-2264.)

   Removing the flush avoids triggering the bug.