capDL-tool: deprecate object back/forward-reference syntax
This syntax was added early on in the design of capDL, but outside of
a few example files it has not been used much. With our upcoming
implementation of allocation via `untypedCovers`, having out-of-order
object references in an untyped could mask genuine allocator bugs,
so we are retiring this feature.