python-capdl-tool: Handle set_mapping on frame cap
By default, the caps that get used to map frames into each component
are kept by the loader. However it is possible to describe in the spec a
mapped cap that gets moved into a target cspace.  This allows components
to perform cache maintenance operations on the mapping or to unmap the
mapping if required. To make this mechanism easier to use, if a cap
passed into get_spec as an existing backing frame has had
set_mapping_deferred called on it, then the spec will call set_mapping
on the cap with the correct mapping container and slot. The loader will
then move the mapped cap into the cspace instead of copying it.