Remove elf attribute from TCB object
This attribute was used by a loader to infer which ELF file to load
into the thread's address space. It is no longer used as the
functionality for loading an ELF is implemented by a fill frames
mechanism for specifying which sections of a frame contains subsections
of an ELF file. This also makes it possible to load specs with sections
of a thread's address space constructed from multiple ELF files.  This
also removes a potential ambiguity where two TCBs with the same VSpace
cap refer to different ELFs.