capdl-loader-app: Fix copy_frame for 64KiB frames
The copy_frame is used for FrameFills and initialising a TCB's stack. It
supports frame sizes that can be mapped into the last-level page table
object, or one level higher. aarch32 has a 64KiB page size that can be
mapped into the lowest level page table.  This means that the copy_frame
symbol for mapping into the last page table needs to be at least this
large.