Remove --code-max-irqs option to capDL-tool
This option was used to synchronise the size of a static array in the C
loader with the number of elements in the mapping array generated by the
tool. It's easier to keep these synchronised by defining the irq array
in the C loader as variable length and allowing the translator to output
an additional num_irqs parameter.