capdl-loader-app: Calculate copy_addr on sys init
The copy_addr variable was statically initialised to round up to
a 16mb alignment on the assumption the kernel only additionally
maps in three extra frames (IPC buffer and bootinfo frames). This
does not account for extended bootinfo frames which could
potentially overlap on the statically calculated 16mb alignment.
Updated copy_addr such that it is calculated with consideration to
the extended bootinfo frames in order to account for all
frames the kernel has mapped in prior.