Add support for Arm GetTrigger and GetTriggerCore
There are additional IRQ control invocations on Arm for specifying the
trigger mode (edge or level) and for specifying the target core. This
adds support to the capDL language to allow specifying IRQs of this
type. Additionally this adds support to the capdl-loader-app for
creating these IRQ handlers and also support in the python-capdl-tool
library for generating Arm specs that have IRQs specifying trigger or
target core attributes.