capdl-loader-app: guard call to seL4_TCB_Configure
If a TCB object isn't provided with a VSpace, CSpace or FaultHandler in
the spec, calls to seL4_TCB_Configure will return an error and none of
the objects will be set. This makes it possible for the loader to create
TCB objects but not configure them with any other objects.