capdl-loader-app: Track number of normal untypes
create_objects() does not use device untypes to create objects, it only
uses normal untypes. When it runs out of normal untypes, it accesses a
part of the untyped array that is empty and tries to allocate objects
from that section resulting in an error from the kernel for trying to
invoke a null capability.

Instead of using this method to halt, we keep track of the number of
normal untypes and stop the loop in create_objects() when we do run out
of untypes.