Fix syntax warnings in capdl_linker.py
Currently, every time the linker is invoked when running from the Docker
setup [0] it prints a syntax warning [1]. This is a bit jarring and
easily avoided.

In this case we are just comparing against a string and don't actually
want to check the identity. Basically, '==' is for comparing values and
'is' is for comparing identities. It just happens to work in Python
either way as strings are immutable.

[0] https://docs.sel4.systems/projects/dockerfiles/
[1]
```
$ ninja
[3/8] Generating spec.cdl
/host/projects/capdl/cdl_utils/capdl_linker.py:113: SyntaxWarning:
    "is" with a literal. Did you mean "=="?
  if args.which is "build_cnode":
/host/projects/capdl/cdl_utils/capdl_linker.py:122: SyntaxWarning:
    "is" with a literal. Did you mean "=="?
  if args.which is "gen_cdl":
[4/8] Generate CPIO archive capdl-loader_archive.o
```

Signed-off-by: David Martin <davidm@brkawy.com>