aarch64: Minimal update for vspace API change
The aarch64 vspace API is changed to only have a single pagetable object
and capability type for all intermediate page table levels. The root
vspace object is still a separate object and capability type.

As part of this change, seL4_ARM_PageGlobalDirectoryObject is no longer
defined for configurations where it was previously an invalid object.
This causes some minor breakages which this commit addresses.

Signed-off-by: Kent McLeod <kent@kry10.com>