Remove seL4_lib dependencies from generated C spec
The capdl_spec.c file is generated by capDL-tool and is compiled into
the capdl-loader-app program to define the system that the loader will
load. The file is supposed to only be a C-language representation of the
capDL spec for a particular system.  Currently, the header files
required to compile this spec file introduces dependencies on a C
library, and libraries from seL4/util_libs and seL4/sel4_libs including
libsel4utils and libutils (and any libraries they depend on).

The spec itself doesn't inherently require any type definitions other
than what's provided by libsel4, and so it shouldn't require these
additional dependencies.

Signed-off-by: Kent McLeod <kent@kry10.com>