capdl-loader-app: Introduce custom error checks
Instead of ZF_LOG* framework functions, we introduce our own layering of
functions that directly output to seL4_DebugPutChar() when the debug
syscall is present and outputs nothing when it isn't. This
implementation allows using a minimal amount of the muslc stdio
implementation for string formatting and bypasses the need for malloc or
any syscalls to be supported.

This will make it easier to refactor capdl-loader-app to be more
freestanding (minimal runtime dependencies) in upcoming commits.

Signed-off-by: Kent McLeod <kent@kry10.com>