capDL-tool: add support for SGISignal caps
Add support for SGISignal caps on Arm. See also RFC-17:
https://sel4.github.io/rfcs/implemented/0170-multikernel-ipi-api.html

SGISignalCaps do not reference objects. IRQHandler caps solve this by
pointing to an artificial IRQ node, but that IRQ node exists at least
in some form in the seL4 implementation. For SGISignalCaps, we chose
to use only cap parameters instead, reflecting their intended use in
multikernel setups where the target IRQ and core are for a different
kernel and have no further representation in the sender kernel apart
from the capability.

Support includes parsing with the cap syntax

sgi_signal (target: <num>, irq: <num>)

and printing to .cdl, XML, C, and Isabelle.

Despite printing support for C, there is no support for SGISignalCap in
the capdl-loader-app or python-capdl-tool yet.

This commit also adds an SGISignalCap example to the AArch32 example for
testing.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>